@@ -9,12 +9,12 @@
         "build": "concurrently -n main,preload,renderer -c blue.bold,yellow.bold,magenta.bold \"npm:build:main\" \"npm:build:preload\" \"npm:build:renderer\" --kill-others-on-fail",
         "build:main": "esbuild src/main/index.ts --platform=node --bundle --outdir=build/main --external:electron --loader:.png=file",
         "build:preload": "esbuild src/preload/index.ts --platform=node --bundle --outdir=build/preload --external:electron",
-        "build:renderer": "vite build",
+        "build:renderer": "vue-tsc --noEmit && vite build",
         "start:dev": "cross-env DEV=true nodemon --watch src --ext ts --ignore 'src/renderer' --exec electron .",
         "start:prod": "electron .",
         "clean": "rimraf build aurora-launcher-win32-x64",
         "prettier": "prettier --config .prettierrc.json --write src",
-        "lint": "eslint . --ext .ts",
+        "lint": "eslint src --ext .ts",
         "lint:fix": "npm run lint -- --fix",
         "pack": "electron-packager build"
     },
@@ -32,36 +32,37 @@
     },
     "homepage": "https://github.com/AuroraTeam/Launcher#readme",
     "devDependencies": {
-        "@sweetalert2/theme-dark": "^5.0.5",
         "@types/semver": "^7.3.8",
         "@types/ws": "^7.4.7",
-        "@typescript-eslint/eslint-plugin": "^4.31.2",
-        "@typescript-eslint/parser": "^4.31.2",
-        "aurora-api": "^0.2.0-beta.2",
-        "concurrently": "^6.2.1",
+        "@typescript-eslint/eslint-plugin": "^5.18.0",
+        "@typescript-eslint/parser": "^5.18.0",
+        "@vitejs/plugin-vue": "^2.3.1",
+        "concurrently": "^7.1.0",
         "cross-env": "^7.0.3",
-        "electron": "^13.4.0",
-        "electron-devtools-installer": "^3.2.0",
         "electron-packager": "^15.4.0",
         "esbuild": "^0.12.29",
-        "eslint": "^7.32.0",
+        "eslint": "^8.12.0",
+        "eslint-plugin-vue": "^8.5.0",
         "nodemon": "^2.0.15",
         "prettier": "^2.4.1",
         "rimraf": "^3.0.2",
         "sass": "^1.47.0",
-        "sweetalert2": "^11.1.7",
-        "typescript": "^4.5.4",
+        "typescript": "^4.6.3",
         "vite": "^2.7.10",
-        "vite-plugin-vue2": "^1.9.2",
-        "vue": "^2.6.14",
-        "vue-router": "^3.5.3",
-        "vue-sweetalert2": "^5.0.2",
-        "vue-template-compiler": "^2.6.14"
+        "vue-tsc": "^0.33.9"
     },
     "dependencies": {
         "@auroralauncher/core": "^0.3.0",
+        "@sweetalert2/theme-dark": "^5.0.10",
+        "aurora-api": "^0.2.0-beta.2",
+        "electron": "^18.0.1",
+        "electron-devtools-installer": "^3.2.0",
         "p-map": "^5.3.0",
         "semver": "^7.3.5",
-        "skinview3d": "^2.2.0"
+        "skinview3d": "^2.2.1",
+        "sweetalert2": "^11.4.8",
+        "vue": "^3.2.31",
+        "vue-router": "^4.0.14",
+        "vue-sweetalert2": "^5.0.2"
     }
 }
