@@ -1,20 +1,23 @@
 module.exports = {
     env: {
+        browser: true,
+        es2021: true,
         node: true,
     },
-    root: true,
-    parser: '@typescript-eslint/parser',
-    plugins: [
-        '@typescript-eslint',
-    ],
     extends: [
         'eslint:recommended',
-        // 'plugin:@typescript-eslint/eslint-recommended',
+        'plugin:vue/essential',
         'plugin:@typescript-eslint/recommended',
     ],
+    parserOptions: {
+        ecmaVersion: 'latest',
+        parser: '@typescript-eslint/parser',
+        sourceType: 'module',
+    },
+    plugins: ['vue', '@typescript-eslint'],
     rules: {
-        "@typescript-eslint/no-explicit-any": "off",
-        "@typescript-eslint/no-var-requires": "off",
-        "@typescript-eslint/ban-types": "off"
-    }
+        '@typescript-eslint/no-explicit-any': 'off',
+        '@typescript-eslint/no-var-requires': 'off',
+        '@typescript-eslint/ban-types': 'off',
+    },
 };
