@@ -1,7 +1,37 @@
+<script setup lang="ts">
+import { ref, onMounted } from 'vue';
+import TitleBar from './TitleBar.vue';
+
+import '../assets/sass/main.sass';
+import Swal from 'sweetalert2';
+
+const inactive = ref(true);
+
+// TODO Доработать
+onMounted(async () => {
+    switch (await launcherAPI.api.getStatus()) {
+        case 'connected':
+            inactive.value = false;
+            break;
+        case 'connecting':
+            break;
+        case 'failure':
+            console.log('apiConnectError');
+            Swal.fire({
+                title: 'Ошибка подключения!',
+                text: 'Сервер не доступен',
+                icon: 'error',
+            });
+    }
+});
+</script>
+
 <template>
     <main>
-        <title-bar />
-        <router-view :class="`main ${inactive ? 'inactive' : ''}`" />
+        <TitleBar />
+        <div class="main">
+            <router-view :class="`${inactive ? 'inactive' : ''}`" />
+        </div>
     </main>
 </template>
 
@@ -9,38 +39,3 @@
 .inactive
     pointer-events: none
 </style>
-
-<script lang="ts">
-import Vue from 'vue';
-import TitleBar from './TitleBar.vue';
-
-import '../assets/sass/main.sass';
-
-export default Vue.extend({
-    components: {
-        TitleBar,
-    },
-    data() {
-        return {
-            inactive: true,
-        };
-    },
-    // TODO Доработать
-    async mounted() {
-        switch (await launcherAPI.api.getStatus()) {
-            case 'connected':
-                this.inactive = false;
-                break;
-            case 'connecting':
-                break;
-            case 'failure':
-                console.log('apiConnectError');
-                this.$swal({
-                    title: 'Ошибка подключения!',
-                    text: 'Сервер не доступен',
-                    icon: 'error',
-                });
-        }
-    },
-});
-</script>
