@@ -1,11 +1,12 @@
 import { defineConfig } from 'vite';
 import { join } from 'path';
-import { createVuePlugin } from 'vite-plugin-vue2';
+import vue from '@vitejs/plugin-vue';
 
 const toDir = (dir: string) => join(__dirname, dir);
 
 export default defineConfig({
     root: toDir('src/renderer'),
+    cacheDir: toDir('node_modules/.vite'),
     base: '',
     build: {
         sourcemap: true,
@@ -14,7 +15,7 @@ export default defineConfig({
         emptyOutDir: true,
         brotliSize: false,
     },
-    plugins: [createVuePlugin()],
+    plugins: [vue()],
     resolve: {
         alias: [
             {
